gui dual export (renderer and main), desktop rpc proven
