+++ examples/desktop/app.js
@@ -1,5 +1,4 @@
 /* global Pear */
-await Pear._ready()
 const { versions, config, updates, wakeups, Window } = Pear
 
 console.log('link', config.link)
