+++ lib/api.js
@@ -9,21 +9,13 @@ class API {
   #refs = 0
   config = null
   argv = global.Bare ? global.Bare.argv : process.argv
-  async _ready () {
-    this.key = this.#ctx.key?.z32 || 'dev'
-    console.log('pre rpc ready')
-    await this.#rpc.ready()
-    console.log('pre rpc config')
-    this.config = await this.#rpc.config()
-    console.log('CONFIG', this.config)
-    this.#ctx.update({ config: this.config })
-    return this
-  }
 
   constructor (client, ctx, onteardown = teardown) {
     this.#rpc = client
     this.#ctx = ctx
     this.#refs = 0
+    this.key = this.#ctx.key?.z32 || 'dev'
+    this.config = ctx.config
     this.#teardowns = new Promise((resolve) => { this.#unloading = resolve })
     onteardown(() => this.#unload())
   }
@@ -70,7 +62,9 @@ class API {
       listener = pattern
       pattern = {}
     }
-    const subscriber = this.#reftrack(this.#rpc.messages(pattern))
+    this.#ref()
+    const subscriber = this.#rpc.messages(pattern)
+    subscriber.on('end', () => this.#unref())
     if (typeof listener === 'function') subscriber.on('data', listener)
     return subscriber
   }
@@ -96,8 +90,4 @@ class API {
   exit = (code) => Bare.exit(code)
 }
 
-if (global.process?.type === 'renderer') {
-  module.exports = require('../gui/api')(API)
-} else {
-  module.exports = API
-}
+module.exports = API
