+++ lib/run.js
@@ -60,10 +60,10 @@ module.exports = async function run ({ rpc, args, key, storage, detached, dir })
         global.process?.exit(1) || global.Bare.exit(1)
       }
 
+      await rpc.ready()
+      ctx.update({ config: await rpc.config() })
+
       const pear = new API(rpc, ctx)
-
-      await pear._ready()
-
       global.Pear = pear
 
       const protocol = new Module.Protocol({
